Add data file of US ZIP codes
This will be helpful in creating an API response that gives a lat/lon
near a given ZIP code.

This work is licensed under creative commons:

```
This means you can use the dump as long as you give credit to geonames
(a link on your website to www.geonames.org is ok) see
http://creativecommons.org/licenses/by/3.0/
```

I'll add a TODO to give them credit.