Extract response formatting from apiClient
This is a pretty sizeable shift but the conditional logic was getting
too dicy in the apiClient.

I separated the responsibility of generating the API.AI response into
its own module that accepts a shelter as a parameter.

I also anticipate a lot more logic in parsing API.AI webhooks, so I
separated that into its own file as well.

This adds eslint as well, which helps improve code quality and catch
simple errors.