Add support for paginating through results
In order for the user to paginate through the list of shelters in the
results, we need to store an array of "seen" shelters in a "context",
which is an API.AI construct for maintaining state.

Since this is the second endpoint, this is also complicated by the
addition of logic around request processing. Now, the responsibilities
look like this:

index.js -- the "server"
* determine which "action" (entity) is being called
* get the response handler for that action
* get the parameters for the response handler
* respond with the status, headers, and body given by calling the
  response handle with the provided parameters

response handler (e.g. `handleLookupShelter`)
* validate its parameters
* make an API call to the Irma API or whatever
* return a status, header, and body

response formatter (called from inside the response handler)
* take the results of the API call, and process them
* return a response body in the format of API.AI's expectation